Feature/person details

+++ app/src/main/java/basaraba/adndrii/movieguide/features/main/movies/MoviesUiEvent.kt
@@ -3,5 +3,5 @@ package basaraba.adndrii.movieguide.features.main.movies
 
 sealed interface MoviesUiEvent {
     object ReloadMoviesScreen : MoviesUiEvent
-    data class ShowMovieDetails(val id: Int) : MoviesUiEvent
+    data class ShowMovieDetails(val id: Long) : MoviesUiEvent
 }
